requirements.txt
anyio==3.6.1
asgiref==3.5.2
cachetools==5.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
dnslib==0.9.19
dnspython==2.2.1
email-validator==1.2.1
google-api-core==2.8.1
google-auth==2.6.6
google-cloud==0.34.0
google-cloud-core==2.3.0
google-cloud-pubsub==2.12.1
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.2
grpc-google-iam-v1==0.12.4
grpcio==1.46.3
grpcio-status==1.46.3
h11==0.13.0
httptools==0.4.0
idna==3.3
itsdangerous==2.1.2
MarkupSafe==2.1.1
orjson==3.7.1
proto-plus==1.20.5
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.1
pymongo==4.1.1
python-dotenv==0.20.0
python-multipart==0.0.5
PythonDNS==0.1
PyYAML==6.0
requests==2.27.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
typing_extensions==4.2.0
ujson==5.3.0
urllib3==1.26.9
